backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "asyncpg==0.31.0",
    "attrs==25.4.0",
    "click==8.3.1",
    "colorama==0.4.6",
    "fastapi[standard]==0.128.0",
    "frozenlist==1.8.0",
    "greenlet==3.3.1",
    "h11==0.16.0",
    "httptools==0.7.1",
    "idna==3.11",
    "multidict==6.7.1",
    "propcache==0.4.1",
    "pwdlib[argon,argon2]>=0.3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pyjwt>=2.11.0",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "redis==7.1.0",
    "ruff==0.15.0",
    "sqlalchemy==2.0.46",
    "starlette==0.50.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uvicorn==0.40.0",
    "watchfiles==1.1.1",
    "websockets==16.0",
    "yarl==1.22.0",
]
